// src/app/interfaces/post.d.ts
// Generated by https://quicktype.io

export interface Post {
  id: number;
  title: string;
  slug: string;
  content: string;
  hit: number;
  category: Category;
  user: User;
  status: string;
  status_color: string;
  createdAt: string;
  publishedAt: string;
  image: Image[];
  tags: number[];
  language: number;
}

export interface Category {
  id: string;
}

export interface Image {
  url: string;
  name: string;
  status: string;
  type: string;
  uid: string;
}

export interface User {
  id: number;
}